refactor(state_store): convert ops:: functions to methods on AppStore/Storage
Move every free function in `state_store::ops` onto the type whose handle
it took: per-app entity I/O (read/write tracking info, component memos,
fn memos, child existence, tombstones, target-state owner, id sequencer,
clear_all, etc.) becomes methods on `AppStore`; the env-level streaming
helpers (`spawn_iter_stable_paths_*`, `list_app_names`) become methods on
`Storage`. The `ops` module and its file are deleted.

This puts the typed I/O on the same handle that gates access to the
underlying database, which is the natural shape for swapping in an
alternate storage backend later: each method becomes a trait method on
the corresponding handle, and call sites don't move.

Also tightens the module boundary:
- `state_store/mod.rs` now declares submodules with `mod` instead of
  `pub mod`, and only re-exports the public types (`AppStore`, `Storage`,
  `StorageSettings`, `AnyTxn`, `ReadTxn`, `WriteTxn`). The `Database`
  alias is now `pub(crate)` since no external caller needs it.
- `AppStore::new` / `AppStore::db` are now `pub(crate)`.
- `Storage::heed_env()` accessor is removed; the streaming methods that
  needed it are now methods on `Storage` itself.
